+++ CHANGELOG.md
@@ -80,6 +80,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   conflict came from (currently this is only supported for conflicts created by
   certain commands).
 
+* `--insert-before` now accepts a revset that resolves to an empty set when
+  used with `--insert-after`. The behavior is similar to `--onto`.
+
 ### Fixed bugs
 
 * Broken symlink on Windows. [#6934](https://github.com/jj-vcs/jj/issues/6934).